fix tests, placeholder checks
