+++ src/lib/placeholders.test.ts
@@ -6,6 +6,13 @@ describe('containsNonArticleCountPlaceholder', () => {
         expect(got).toBe(true);
     });
 
+    it('returns true if string contains two placeholders (that is not %%ARTICLE_COUNT%%)', () => {
+        const got = containsNonArticleCountPlaceholder(
+            "You've read %%ARTICLE_COUNT%%. Support the Guardian from as little as %%CURRENCY_SYMBOL%%1 – and it only takes a minute. Thank you.",
+        );
+        expect(got).toBe(true);
+    });
+
     it('returns false if string contains an article count placeholder', () => {
         const got = containsNonArticleCountPlaceholder('apple %%ARTICLE_COUNT%%SDF%%');
         expect(got).toBe(false);
